requirements.txt
langchain>=0.2.14
langchain-community>=0.2.14
langchain-openai>=0.2.14
pydantic>=2.0.1
PyYAML>=6.0
loguru>=0.7.3
rouge-score>=0.1.2
torch>=2.0.0